pyproject.toml
[project]
dynamic = ["version"]
name = "VirtIAC"
description = "Infrastructure-as-code for LibVirt"
keywords = ["libvirt"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience  :: Development",
    "Programming Language  :: Python",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "libvirt-python>=11.0.0",
    "xdg-base-dirs>=6.0.2"
]
requires-python = ">=3.11"
authors = [
    {"name" = "Daniele Penazzo", "email" = "penazarea@altervista.org"},
]
maintainers = [
    {"name" = "Daniele Penazzo", "email" = "penazarea@altervista.org"},
]

[project.optional-dependencies]
dev = [
    "ipdb"
]
